# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API RESTful - NoteSwift",
    "version": "1.0.0",
...
    "license": {
      "name": "MIT License",
      "url": "https://github.com/ViniciusGR797/noteswift-api/blob/main/LICENSE"
    }
  },
  "components": {
    "securitySchemes": {
      "jwt": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Para autenticar, coloque o token JWT no formato **Bearer _token_**."
      }
    },
    "schemas": {
      "Library": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Folder"
        }
      },
      "Folder": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "is_default",
          "color",
          "order",
          "notes"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Identificador único da pasta",
            "example": "6123456789abcdef01234567"
          },
          "name": {
            "type": "string",
            "description": "Nome da pasta",
            "example": "Nome da pasta"
          },
          "is_default": {
            "type": "string",
            "description": "Marcação da pasta default (não pode ser apagada)",
            "example": true
          },
          "color": {
            "type": "string",
            "description": "Cor da pasta",
            "example": "#FF0000"
          },
          "order": {
            "type": "integer",
            "description": "Ordem da pasta",
            "example": 1
          },
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Note"
            }
          }
        }
      },
      "FolderCreate": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da pasta",
            "example": "Nome da pasta"
          }
        }
      },
      "FolderUpdate": {
        "type": "object",
        "required": [
          "name",
          "color"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da pasta",
            "example": "Nome da pasta"
          },
          "color": {
            "type": "string",
            "description": "Cor da pasta",
            "example": "#FF0000"
          }
        }
      },
...
      "Note": {
        "type": "object",
        "required": [
          "_id",
          "title",
          "body",
          "style",
          "trashed",
          "update_at"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Identificador único da nota",
            "example": "6123456789abcdef01234567"
          },
          "title": {
            "type": "string",
            "description": "Título da nota",
            "example": "Título da nota"
          },
          "body": {
            "type": "string",
            "description": "Corpo da nota",
            "example": "Corpo da nota"
          },
          "style": {
            "type": "string",
            "description": "Estilo da nota",
            "example": "Estilo da nota"
          },
          "trashed": {
            "type": "boolean",
            "description": "Indica se a nota está na lixeira",
            "example": true
          },
          "deleted_date": {
            "type": "string",
            "description": "Data de exclusão da nota",
            "example": "2022-01-01 10:30:00"
          },
          "update_at": {
            "type": "string",
            "description": "Data de atualização da nota",
            "example": "2022-01-01 10:30:00"
          }
        }
      },
...
      "DeleteSuccess": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Mensagem de sucesso ao excluir",
            "example": "Excluído com sucesso"
          }
        }
      },
...
      "BadRequest": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Requisição inválida. Alguns parâmetros podem estar faltando ou serem inválidos",
            "example": "Alguns parâmetros podem estar faltando ou serem inválidos"
          }
        }
      },
      "Unauthorized": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Não autorizado. O token JWT não foi fornecido ou é inválido",
            "example": "O token JWT não foi fornecido ou é inválido"
          }
        }
      },
      "Forbidden": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Proibido. O usuário não tem permissão para acessar o recurso solicitado",
            "example": "Não tem permissão para acessar o recurso solicitado"
          }
        }
      },
      "NotFound": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Nenhum dado encontrado. O recurso solicitado não existe",
            "example": "Nenhum dado encontrado"
          }
        }
      },
      "InternalServerError": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string",
            "description": "Erro interno ao manipular dados no serviço",
            "example": "Erro interno do servidor"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "jwt": []
    }
  ],
  "paths": {
...
    "/folders": {
      "get": {
        "summary": "Vizualiza pasta default",
        "description": "Retorna as informações da pasta default",
        "tags": [
          "Folder"
        ],
        "operationId": "get_default_folder",
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Folder"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria uma nova pasta",
        "description": "Cria uma nova pasta com base nos dados fornecidos no corpo da requisição.",
        "tags": [
          "Folder"
        ],
        "operationId": "create_folder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Folder"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/folders/{folder_id}": {
      "parameters": [
        {
          "name": "folder_id",
          "in": "path",
          "required": true,
          "description": "ID da pasta que será deletada",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Visualiza pasta pelo _id",
        "description": "Retorna as informações da pasta pelo _id",
        "tags": [
          "Folder"
        ],
        "operationId": "get_folder_by_id",
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Folder"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Atualiza pasta",
        "description": "Atualiza as informações da pasta desejada",
        "tags": [
          "Folder"
        ],
        "operationId": "update_folder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderUpdate"
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Folder"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove pasta",
        "description": "Apaga pasta, lembrando que a pasta default não pode ser apagada",
        "tags": [
          "Folder"
        ],
        "operationId": "delete_folder",
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSuccess"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/folders/name/{folder_name}": {
      "parameters": [
        {
          "name": "folder_name",
          "in": "path",
          "required": true,
          "description": "Nome da pasta que será visualizada",
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "Lista pastas pelo name",
        "description": "Retorna as informações da pastas pelo name",
        "tags": [
          "Folder"
        ],
        "operationId": "get_folder_by_name",
        "security": [
          {
            "jwt": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Library"
                }
              }
            }
          },
          "400": {
            "description": "BadRequest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "500": {
            "description": "InternalServerError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}